update GuLogShippingPolicy to bring it's own parameter
We've established best practice for accounts/stacks to define the kinesis stream to send logs to to be a SSM Parameter called `/account/services/logging.stream.name`.

This updates `GuLogShippingPolicy` to add the parameter to the stack within that construct. The construct is bringing its own resources with it, rather than being passed in as props. This makes it more self contained.