update the GH Action trigger
It looks like releases and tags are different concepts in GitHub, even though, confusingly, tags show up on https://github.com/guardian/cdk/releases.

It looks like our release process is currently only creating tags and NOT creating releases. As a result, this change tweaks the GH Action to update docs on _tag_ creation.

We'll likely add semantic-release to this library, which does in fact create releases, but this is a little far away.