chore: further automate release process
We use `np` to orchestrate the publication of a module to NPM. One of the steps `np` performs is a commit to edit `package.json`'s `version`.

As we have branch protection enabled for `main`, we have to raise a PR to get this commit in. Automate this process in the release script as it's pretty boring!

This change assumes one has the GitHub CLI installed.