chore: simplify CI to run a single script
This change updates the GHA for CI to run a single script in the repo. Whilst this results in running steps in sequence, and thus a potential to increase runtime, it makes things easier to reason about.

This becomes particularly apparent when needing to add a GHA job that depends on others completing, for example post-release-action.