refactor: create a simpler interface for defining security group rules (#319)
To define an ingress or egress rule, we now provide a SecurityGroupAccessRule.

SecurityGroupAccessRule has three mandatory fields:
- range which is the CIDR range
- port which is the port number and protocol (tcp by default)
- description to provide some useful information to the reader viewing the security group in the AWS console

BREAKING CHANGE:
`CidrIngress` and `CidrEgress` have been removed in favour of a single `SecurityGroupAccessRule`. This change forces a description to be provided for every rule as it makes it easier to reason about in the AWS console.

It also changes the type of port to `number | Port` - if you provide a number the protocol will default to TCP as this is the most common use-case. This should provide a simpler API.

`transformToCidrIngress` is renamed to `transformToSecurityGroupAccessRule` as it's not longer tied to ingress rules.

Port 22 is completely forbidden as SSH over SSM is preferred.