refactor(GuLogShippingPolicy): make `GuLogShippingPolicy` a singleton
This policy is now a singleton as we only ever want one per stack. If there are multiple roles defined in this stack that require log shipping permissions, a single policy will be defined and attached to the roles. This DRYness helps keep within the file size limits of cloudformation.