fix: simplify props for `GuClassicLoadBalancer`
Removes the `propertiesToRemove` prop. It looks like we were using this to set the scheme on the load balancer, however this is already being driven by the `internetFacing` prop provided by AWS.

For a scheme of "internal", set `internetFacing` to `false`. For a scheme of "internet-facing", set `internetFacing` to `true`.

See:
- https://github.com/aws/aws-cdk/blob/fb6512314db1b11fc608cd62753090684ad0d3c4/packages/%40aws-cdk/aws-elasticloadbalancing/lib/load-balancer.ts#L249
- https://docs.aws.amazon.com/AWSCloudFormation/latest/UserGuide/aws-properties-ec2-elb.html#cfn-ec2-elb-scheme