chore: Add custom Jest matcher to simplify logicalId testing
It is fairly common to want to test the logicalId of a resource. This can be tricky as:
  - constructs that have an `AppIdentity` will suffix the `id` with the app value
  - not every construct supports static logicalIds and will always be auto-generated

Rather than passing in the `id` with a combination of `migratedFromCloudFormation` and `existingLogicalId` to get a static/deterministic logicalId, we can pass a regex to the new `toHaveResourceOfTypeAndLogicalId` matcher, asserting the stack has a resource of type x with logicalId y.