fix: Adopt new logicalId override logic across the constructs
In #364 we simplified and encapsulated the logic to override a construct's logicalId.

This change adopts the new logic into the constructs.

As a consequence to this a number of tests have been moved to using the auto-generated logicalId.
This reduces the variables in play in our tests and therefore simplifies them.