chore(deps): Keep aws-cdk dependencies in integration-test up to date
We perform a check at CI time to ensure the version of aws-cdk libraries we're using are within 5 versions of the latest available.

If we're out of date, CI fails with instructions to run `./script/update-aws-cdk`. This script uses ncu (npm-check-updates) to update the aws-cdk dependencies.

This change updates the call to ncu to:
  - Configure ncu to update the "aws-cdk" dependency in addition to @aws-cdk/*
  - Add the --deep flag to "Run recursively in current working directory."

See:
  - https://www.npmjs.com/package/npm-check-updates
  - #192