chore(deps): configure dependabot for the integration-test project
The integration-test directory holds an... integration test.

It is a standalone project that:
  - installs @guardian/cdk from a relative path
  - defines a stack
  - synths the stack

This is to observe any regressions in the packaging of @guardian/cdk.

This change adds dependabot to the integration test project, as it is a project after all and its dependencies need to kept up to date.

See:
  - #367