fix: Update `GuApplicationListener` with revised logicalId logic
In #364 we placed the logic of overriding a construct's logicalId into a single place.

In this change, we're updating the `GuApplicationListener` construct to adopt the new logic. As of #418 it's as simple as using the `GuStatefulMigratableConstruct` mixin!

The logicalId logic currently in `GuApplicationListener` is very complicated with multiple paths to overriding the logicalId.

This is massively simplified, as evidenced in the updated tests.