fix: Log messages as YAML comments
When a stack is synthesized, AWS CDK will:
  - print the resulting CloudFormation template as YAML to the console
  - save the resulting CloudFormation template as JSON to disk

The AWS CDK library does not support synthesizing a YAML template to disk. The only way to achieve this is is to redirect the result of `cdk synth` to disk.

For example:

```console
cdk synth > cloudformation.yaml
```

Since #364 we have started printing important messages to the console. This breaks the saving YAML to disk trick as the file on disk is not valid YAML.

For example:

```yaml
GuStack has 'migratedFromCloudFormation' set to false. MyDatabase is a stateful construct, it's logicalId will be auto-generated and AWS will create a new resource.
Parameters:
  Stage:
    Type: String
    Default: CODE
    AllowedValues:
      - CODE
      - PROD
    Description: Stage name
Resources:
  MyDatabaseA1B2C3D4:
    Type: AWS::RDS::DBInstance
```

For this reason, prefix console messages with "# ". This results in `cloudformation.yaml` still being valid YAML.

For example:

```yaml
\# GuStack has 'migratedFromCloudFormation' set to false. MyDatabase is a stateful construct, it's logicalId will be auto-generated and AWS will create a new resource.
Parameters:
  Stage:
    Type: String
    Default: CODE
    AllowedValues:
      - CODE
      - PROD
    Description: Stage name
Resources:
  MyDatabaseA1B2C3D4:
    Type: AWS::RDS::DBInstance
```