feat: Simplify props for `GuLambdaFunction`
Simplifies the props needed to create a `GuLambdaFunction` by removing the bucket key in favour of using the `GuDistributionBucketParameter` singleton.

This brings `GuLambdaFunction` in line with our EC2 constructs and patterns.

`GuDistributionBucketParameter` adds an SSM parameter to a stack with a default path of `/account/services/artifact.bucket`.

BREAKING CHANGE: Shape change to the props for `GuLambdaFunction`

The `code.bucket` field has been removed.