chore: Fix "duplicate manual mock found" message
In #448 we started including the mocks in the published package.

Since then the following message is appearing when running tests,
both locally and in CI:

```log
jest-haste-map: duplicate manual mock found: library-info
  The following files share their name; please delete one of them:
    * <rootDir>/lib/constants/__mocks__/library-info.js
    * <rootDir>/src/constants/__mocks__/library-info.ts
```

This change configures Jest to ignore the `lib` directory,
which should quieten this message.

This should not affect the published package.

See:
  - https://jestjs.io/docs/configuration#modulepathignorepatterns-arraystring
  - https://github.com/facebook/jest/issues/6801#issuecomment-712951064
  - #448