fix: Add `AppIdentity` to `GuClassicLoadBalancer`
Adding `AppIdentity` to `GuClassicLoadBalancer` to support a multi-app stack.

This means the auto-generated logicalId gets suffixed with the app string
and the resource also gets the App tag.

This isn't considered a breaking change because `GuClassicLoadBalancer`
only gets used when migrating a stack and not for new stacks.

That is, we can expect `existingLogicalId` will always get set,
which will prevent the logicalId from being auto-generated.