chore: Check item is a directory first
We crawl some directories to include in the TypeDoc documentation site.

This fixes this warning in the build log:

```
Warning: Provided entry point src/utils/logger.ts/index.ts does not exist and will not be included in the docs.
```

It looks like we need to explicitly check if an item returned from `readdirSync` is a directory.