feat: Simplify props for `GuLambdaFunction`
Currently the props to `GuLambdaFunction` are a little confusing.

The props currently have an optional `rules` and `apis` property.

This means it is possible to create a lambda that:
  - Sits there with no event source
  - Runs on a cron
  - Runs behind API Gateway
  - Runs on a cron and behind API Gateway

We never want the first and last option to be possible.

This change moves the props for a cron lambda into the `GuScheduledLambda` pattern for a cleaner API.

BREAKING CHANGE: The shape of the `GuScheduledLambda` pattern has changed.