chore(guparameter): add a sensible default to GuParameters when using SSM
At the moment, there is no indication in the UI that the parameter you are filling in is an SSM
parameter. Users may remember to add a sensible default or a hint in the description, but they also
may not. This change ensures that, when not provided explicitly, users are hinted that the parameter
is from SSM as well as what the ideal SSM path should be.