fix: Support singletons across multiple stacks
A singleton is something that will be instantiated once and once only.

Or rather, a singleton construct is one which will be instantiated once and once only in a `Stack`.

This change fixes a bug where a singleton could not be used within two `Stack`s that get instantiated using the same `App` instance.

Something like this was not possible as there would be a cyclic reference:

```typescript
class MyStack extends GuStack {
  constructor() {
    GuDistributionBucketParameter.getInstance(this);
  }
}

class MyOtherStack extends GuStack {
  constructor() {
    GuDistributionBucketParameter.getInstance(this);
  }
}

const app = new App();

const myStack = new MyStack(app, "MyStack", { stack: "deploy" });
const myOtherStack = new MyOtherStack(app, "MyOtherStack", { stack: "deploy" });

myStack.synth();
myOtherStack.synth();
```

We'd get an error like this:

> Error: 'MyStack' depends on 'MyOtherStack'. Adding this dependency (MyOtherStack -> MyStack/DistributionBucketName.Ref) would create a cyclic reference.

This change checks if a singleton exists on the `Stack`, rather than the `App`, to fix this issue.

A test is added to the `GuDistributionBucketParameter` construct to demonstrate this.