feat: Add new tag for repository info
Adds a new `gu:repo` tag to the GuStack and any resource added to it.

The purpose of this tag is to provide information on where to find  an
application.

The value of the tag comes from one of two places:
  1. CDK context
  2. Git config

If we cannot find the repository name from either, we log a message.

A Git URL can be in two main formats:
  1. https, for example `https://github.com/guardian/cdk.git`
  2. ssh, for example `git@github.com:guardian/cdk.git`

Use `git-url-parse` to handle the parsing of a Git URL and present
it in the form `owner/name`.

For example, `guardian/cdk` or `guardian/prism`.

See:
  - https://docs.aws.amazon.com/cdk/latest/guide/context.html