chore: Add lockfile to `integration-test` project
We previously had issues with lockfiles in CI for the `integration-test` project.
This caused a few issues and led to the changes in #569.

This change reverts those commits and adds a lockfile.
That is, I'm not sure what the previous issues were, but they don't seem to
exist anymore and we can use lockfiles in CI and gain deterministic builds.