chore!: remove GuPublicInternetAccessSecurityGroup (#548)
BREAKING CHANGE: `GuPublicInternetAccessSecurityGroup` has been removed, as AWS automatically creates a security group which is almost identical to this by default when creating an Application Load Balancer. If you previously created an instance of `GuPublicInternetAccessSecurityGroup` and you need to retain the same custom security group, pass the appropriate ingresses and description to `GuSecurityGroup` instead of using the `GuPublicInternetAccessSecurityGroup` helper.