feat: Add policy to allow writing to Anghammarad
Create a policy that allows `sns:Publish` to the Anghammarad SNS topic.

The Anghammarad SNS topic is used across multiple stacks,
so we can promote it to an account wide ("account/services") parameter store location.

We then wrap this account wide parameter into the `AnghammaradTopicParameter` singleton,
which the policy will add when necessary.