feat!: Make `GuStageParameter` a singleton
Converts the `GuStageParameter` construct into a singleton to simplify
references to it across constructs.

Before, we had to something like:

```typescript
const stage = this.getParam<GuStageParameter>("Stage");
```

With this change, this becomes:

```typescript
GuStageParameter.getInstance(this).valueAsString
```

BREAKING CHANGE: `GuStageParameter` is now a singleton

Making `GuStageParameter` a singleton simplifies access to it.
Note, `this.stage` on the `GuStack` is still possible.