chore: Enable beta releases to NPM
Whilst it is possible to install this package from GitHub,
it can be useful to install from NPM as the published package structure varies slightly.

With this change, we will be able to publish to the `beta` channel by update the `beta` branch.
This will enable us to to install our experimental changes from NPM with:

```
npm install @guardian/cdk@beta
```