feat: Add `account-readiness` CLI command
The `account-readiness` CLI command will check for the existence of SSM parameters in an account.

Note: It does not check if the value is correct, just that it exists.
Although `LoggingStreamName` is described as "Name (not ARN)",
this command will not validate the value in parameter store against this requirement.

The command supports the `--profile` and `--region` flags, similar to AWS's CLI.