feat: Perform VPC analysis in account-readiness
GuCDK uses values in SSM parameter store to know into which VPC resources should be launched.

If the parameters do not exist, GuCDK will produce a template that cannot be executed.

This change adds some VPC analysis to the `account-readiness` CLI command. It:
  - Finds all VPCs in the AWS region
  - Determines if the VPC is in use
  - Expects three SSM parameters to exist for each in use VPC:
    1. ^/account/vpc/([A-z0-9.-_])+/id$
    2. ^/account/vpc/([A-z0-9.-_])+/subnets/private$
    3. ^/account/vpc/([A-z0-9.-_])+/subnets/public$

Typically, an account has one in use VPC.