cd: Perform preamble checks at CD time too
Our current CI process checks that the version of AWS CDK doesn't fall too far behind the latest release.

Our current CD process will parse commits to determine if a new release is needed.
If a release is needed, a PR is made soon after to bump the version number in `package.json`.
This PR is then merged by a robot once CI passes.

There is a little gap in our process. Let's say:
  - PR raised and the aws check passes
  - A new release to AWS CDK has been made
  - PR merged
  - A new GuCDK release made
  - The PR to bump the version number in `package.json` fails CI checks and cannot be merged automatically

This leaves us in a slightly broken state.

This change performs the aws check during CD too to help resolve this. That is:
  - PR raised and the aws check passes
  - A new release to AWS CDK has been made
  - PR merged
  - CD on `main` fails and no new release is made