ci: Move AWS CDK check into it's own GHA job
Currently CI will fail when the version of AWS CDK libraries being used are over 5 releases out of date.
This forces us to keep somewhat up to date with releases, however it does introduce an odd situation:
  1. Raise PR
  2. CI passes on PR
  3. PR gets approved
  4. New AWS CDK released
  5. PR merged
  6. Semantic Release creates a new release
  7. Robot that keeps `version` in `package.json` up to date raises PR
  8. Robot PR fails CI and cannot be automatically merged

When this happens we usually update the branch of the robot's PR to remove the AWS CDK check,
allowing CI to pass.
This isn't great as we put the check back in the next PR. This is noise.

This change moves the AWS CDK check into a discrete job in GHA.
This should mean we can resolve the above solution by updating the branch protection rules,
allowing the AWS CDK check to fail but still require the tests to pass.

This yields less noise in the repository.