chore: Forgo keeping `version` in `package.json` up to date
Currently when a new release is made (via Semantic Release), a robot will raise and merge a PR to update `version` in `package.json`.

This is handy as we can browse the file to see what the latest version is.
I don't think it's necessary for the release process - Semantic Release gets the latest version number from npm.

It comes at a slight cost, however:
  - A bit of PR noise. We're not expected to interacted with the robot's PRs, but the `CODEOWNERS` rules mean we get added as reviewers.
  - An odd scenario where our CI checks don't pass and the robot's PR cannot be automatically merged. See #998.

This change moves to a simpler model of keeping `version` in `package.json` completly static between releases.
If one would like to find the number of the latest version, npm or GitHub releases can be used.