fix: `GuEcsTask` uses `stack` and `stage` from parent `GuStack` for consistency
We do not need to specify `stack` and `stage` at the component/pattern level as these properties are available on `GuStack`.
Using the `GuStack` values also increases the type-safety as `stage` is _not_ a `string`.

This change also refactors to use destructing of the `props` object, which yields simpler code for default values.