Add stage refactoring POC
The main aim was to make constructs and patterns ignorant of
stages. This achieves a few things:

* keeps them simple (LOC and also readability/complexity)
* ensures our codebase can handle unusual stage requirements
  without significant refactoring
* simplifies the INFRA case (it requires no special code at all)
* makes it easier to e.g. use `cdk deploy` in the future

The downside is that constructs would become less opionnated on
which props should be stage-specific. Generally, I think this is
fine, but there might be cases where e.g. if one prop is customised
by stage we would like to ensure another (related) prop is too.