feat: Refactor account-readiness
* remove /account/vpc/default/* check as no longer required
* move default VPC reference in parameters check to SSM readiness check
* remove 3-params-per-VPC check as noisy
* remove default-VPC-in-use check as noisy

By 'noisy', I mean that these are not actually blockers to getting
started with CDK and most accounts fail on one or both. For example,
deployTools fails on these even though we are happily using CDK to
deploy our core apps.

I've also introduced a 'Report' type to make standard reporting
a bit easier.

SSM readiness output has been modified slightly as a result, but
not in a major way.