fix: Remove YAML friendly log messages
Following #1151 where we advise against redirecting the output of `cdk synth` to a `cfn.yaml` file,
this change removes some code that was initially introduced to enable this workflow.

That is, this change further encourages the use of the JSON files produced in `cdk.out`,
as any `cfn.yaml` file produced by redirecting to disk might now be invalid.