fix: add withoutImdvs2 to props of GuAutoScalingGroup and GuEc2App
We're adding this prop to allow consumers of the GuAutoScalingGroup construct
and GuEc2App pattern to be able to explicitly opt out of IMDSv2. Previously,
requireImdsv2 was hardcoded to true inside of GuAutoScalingGroup. It's best practice
to use IMDSv2, but it's good to provide an escape hatch for applications that
aren't ready to adopt it.

NB: GuAutoScalingGroup defaults the value to false, so existing consumers will
be unaffected by this change.