chore(deps): Configure Dependabot to ignore `constructs`
Users of `@guardian/cdk` must be on the exact version of the `aws-cdk` and `construct` dependencies,
else they see issues like this in CI (source https://github.com/guardian/cdk-playground/pull/192):

```console
    lib/cdk-playground.ts:17:52 - error TS2345: Argument of type 'this' is not assignable to parameter of type 'IConstruct'.
      Type 'CdkPlayground' is not assignable to type 'IConstruct'.
        Types of property 'node' are incompatible.
          Type 'import("/home/runner/work/cdk-playground/cdk-playground/cdk/node_modules/constructs/lib/construct").Node' is not assignable to type 'import("/home/runner/work/cdk-playground/cdk-playground/cdk/node_modules/@guardian/cdk/node_modules/constructs/lib/construct").Node'.
            Types have separate declarations of a private property 'host'.

    17     AppIdentity.taggedConstruct(CdkPlayground.app, this);
```

The above was caused by merging https://github.com/guardian/cdk/pull/1195 w/out creating a new major version release,
and it consequently being included in the 41.0.1 patch.

In this change, we remove management of `constructs` from Dependabot in favour of bumping the version ourselves (and creating a new major release when doing so).