feat(tagging): Tag resources with build number
The build number is one (/the only?) way to know which version of the infrastructure definition is running,
and is useful information when triaging incidents.

In this change we look up the build number from environment variables:
  - `BUILD_NUMBER`
  - `GITHUB_RUN_NUMBER`

If not found, we fall back to `"unknown"`.

Coupled with https://github.com/guardian/devx-logs and https://github.com/guardian/cloudwatch-logs-management,
this tag will get automatically added as markers to logs shipped to the ELK stack.