feat: Add `overrideLogicalId` function to `GuStack`
With an aim to remove the `GuStatefulMigratableConstruct` mixin and the `GuMigratingStack` interface,
this change adds a helper function to `GuStack` to override logical IDs.

This will mean _any_ CDK construct can override it's logical ID, not just those with GuCDK flavours.

As an example, `GuWazuhAccess` has been updated and the tests still pass!