fix(GuVpc): Configure AZs for eu-west-1 region, so you don't have to
The number of AZs in eu-west-1 is fairly static (at least it hasn't changed in over a decade).

In this change, we programmatically set the context for the `GuVpc` construct
so that users don't have to commit the, slightly magical, `cdk.context.json` file.

If `GuVpc` is used in a different region, the `cdk.context.json` file still needs to be added to VCS by users.