ci: Update AWS CDK libaries on a schedule
Inspired by https://github.com/cdk-dev/bump-cdk-action, this change introduces a scheduled job to update:
 - aws-cdk
 - aws-cdk-libs
 - constructs

As bump-cdk-action describes:

> Because of the way CDK is compiled, all peer dependencies must be the same version.
> This causes a maintenance headache of not only bumping the versions,
> but also making sure they go together.

If aws-cdk and aws-cdk-libs are not on the same version, you're greeted with this error upon synth:

```console
This CDK CLI is not compatible with the CDK library used by your application. Please upgrade the CLI to the latest version.
(Cloud assembly schema version mismatch: Maximum schema version supported is 20.0.0, but found 21.0.0)
```

Ideally we'd use Dependabot to achieve this, however grouped dependencies are not supported.
See https://github.com/dependabot/dependabot-core/issues/1296.