ci: Remove AWS CDK version check from CI
Reduce friction of raising a PR by the AWS CDK version check.

First introduced in #192, these checks were added to force us to keep up to date with the rapid release cadence.

However as we've recently slowed down development in the repository, this requirement is causing friction for contributors.

Let's remove it in favour of some automation.