feat(GuS3OriginBucket): Add `GuS3OriginBucket` construct
The `GuS3OriginBucket` is to create a bucket destined to be an origin for a static site,
whether that be a Fastly origin, CloudFront origin, etc.

In addition to the bucket, it will also create an SSM Parameter on a known path,
whose value is the bucket name.

This Parameter should then be used by dependant services,
as the bucket name is considered private data.