fix: Switch from our home baked SSM policy to the current managed policy
Once upon a time the recommended managed policy allowed full access to all S3 buckets but this craziness is no longer the case and the managed policy is treated with care. We should switch to this, at least in part, because it has some extra permissions needed for AWS Inspector (v2) and new required SSM permissions will be added for us in the future.