feat: Add DeployToolsAccountId named SSM parameter
The initial motivation is to reference a shared Lambda Layer that
lives in the Deploy Tools account. But I suspect this will be
generally useful given we often locate shared DevX resources here.

An alternative is:

https://github.com/guardian/private-infrastructure-config

But am reluctant to add this as a dependency to all CDK users as
it contains private info and requires private repo access.