fix(email): Remove sender restriction
Removes the restriction for an email to be sent from the domain `theguardian.com`.

There are a few services in the Journalism stream using the `gutools.co.uk` domain,
so this restriction was overly tight.

Additionally, "shift left" by making the email sender a compile time prop,
rather than a CFN Parameter, which is only known at runtime.