chore: Add workflow to combine PRs
Use https://github.com/github/combine-prs to combine Dependabot PRs
to make them more manageable.

The resulting PR will have a `fix` prefix, meaning it'll create a patch release on merge.
This is so that any vulnerability fixes in the bumped dependencies go out early,
as opposed to waiting for a feature change PR.