feat(riff-raff.yaml): Support user-added deployment types
This change makes the `riffRaffYaml` property of `RiffRaffYamlFileExperimental` public.
This means we can supply additional deployment types, if needed.

The typical use-case is adding an `aws-s3` deployment type,
which `RiffRaffYamlFileExperimental` does not currently support.

No validation is performed[^1] on the parameters you provide,
so you might get deployment errors due to an invalid `riff-raff.yaml` file.
For example, if an unrecognised parameter or deployment type is provided,
an invalid `riff-raff.yaml` file will be created.

Usage of this should be uncommon.

[^1]: There is some typing, but ultimately we're dealing with the `string` type.