feat: Add 4xx alarms as an optional parameter to ec2 base alarm config
Being alerted to high volumes of 404 requests is arguably a common pattern that is useful to extend to CDK. We've added this to another project where it would have allowed us to catch a production issue, so contributing upstream to gauge if useful for others.

This change adds an optional http4xxAlarm to the `Alarms` type on the `GuEc2AppProps` interface, as it is perhaps not as critical to consider as the 5xx events, but still hopefully a useful affordance.